Remove 'm' parameter from HasClient
This parameter was presumably added to the typeclass so that additional constraints
(`RunClient`, `RunStreamingClient`) could be put on the underlying monad.

It proves to be a hindrance when implementing the `HasClient` class for
`GApi api`. The reason seems to be that, since `hoistClientMonad` does _not_
require any additional constraints on the client monads (and rightly so), GHC cannot
prove that client types actually exists and thus fails to verify
necessary type equalities.

The solution chosen here is to specify the constraints as an associated type family
using ConstraintKinds, and use them in the definition of
`clientWithRoute`. The monad parameter in the `HasClient` class can
hence be removed, and GHC can reason more easily about the shape of
generic clients.

In spite of the machinery involved, I think this change conceptually makes sense: there is no reason
that `HasClient` should depend on a specific monad. It makes things more
symmetric with `HasServer`.